Fix Arena 10K Oracle experiment pipeline issues
- Clean up duplicate entries in p0_replies.jsonl (16 failed + 16 successful → 32 unique)
- Fix provider registry missing get_provider function and attributes
- Add cje/providers.py module for backward compatibility
- Fix judge scoring float conversion issue (was truncating 0.6 to 0)
- Update provider registry to remove legacy code and add missing attributes
- Document judge configuration: llama4-scout-instruct-basic at T=0.0 with comprehensive_judge template
- Successfully generate judge scores for all 32 responses (avg: 0.83, range: 0.5-1.0)

+++ experiments/arena_10k_oracle/scripts/03_add_judge_scores.py
@@ -129,9 +129,9 @@ def score_responses_with_judge(
                 if isinstance(score_result, dict):
                     score = score_result.get("score", None)
                 else:
-                    # Try to parse as integer if string
+                    # Keep as float or try to parse
                     try:
-                        score = int(score_result)
+                        score = float(score_result)
                     except:
                         console.print(
                             f"⚠️  [yellow]Failed to parse score: {score_result}[/yellow]"
